Fix Filename Errors in Env Override Test Case
The test case for checking the env override behavior between the
`DEFAULT_GRAMMLOG_LEVEL` and the `GRAMMLOG_LEVEL` creates a random
filename to ensure it a log file created from another test case doesn't
interfere with the results of the test. But I was using a stupid method
of generating the name. For some reason, I stringified random bytes and
used that result directly as the filename. I don't know why I thought
that would work as a filename in all cases, but regardless, the Windows
CI runner failed last night when I pushed the initial upload, and I
assumed that was because windows makes everything difficult (which it
normally does), but this one was all on me.

I have updated the test case to use 8 random alpha characters instead of
random bytes, and I also reenabled the Windows runner, so hopefully
everything will pass now. I don't have a local windows machine to test
this on though.

Note:
    I may eventually start using a temporary directory for the log files
    in the test suite, but I prefer keeping the test environment as
    close to production as possible, so for the time being, I will
    continue to create actual log files in the test suite and simply
    gitignore them. Once I have a complete regression suite with test
    cases for all of the invariants of the library, I will probably look
    into making things more efficient.