Remove Main-Push from CI Run Triggers
The CI workflow was running on push and PR for both dev and main.

This is good to ensure that code that gets pushed directly to main is
checked, but this should not happen in practice, and the trigger on push
to main causes the CI workflow to run twice when merging from dev. Once
when the PR is created, and again once it is merged into main. This is
annoying since I normally create a release right after merging a PR, and
the windows runner takes over 2 minutes to complete.

I removed the trigger on push to main to avoid this speedbump when
preparing a release.