Rebuild Documentation for Version 1.0.1
I forgot to rebuild the docs site when releasing the hotfix.

The documentation hasn't changed other than the version and some
comments in the source code that is viewable via the *view source*
toggles in the api docs generated by pdoc.

This doesn't really affect anything, but it's better to keep things
consistent, so I am pushing this to ensure that the public docs site
reflects the current latest release of the source code.