Fix Timestamp Handling in Async Logging Functions
The `async_*` functions were broken because they were passing in the
public log functions for the levels info through critical instead of the
private `_{level}` functions that take the extra `timestamp_override`
argument. This was causing a runtime failure when calling those
functions due to the argument mismatch.

I added a test case to ensure that the timestamps were being correctly
handled in the async functions. Testing async code can be complicated,
so I'll provide an explanation of this test case here.

The test case constructs a couple of async tasks. One of them calls the
`async_*` log function to log a
message, and the other calls `time.sleep(0.5)` and then returns a
timestamp after the sleep. We then read the written log message and
assert that it is at least `0.4` less than the timestamp returned after
the sleep. This indirection is necessary because the actual write to
disk happens in the background in a separate task that is controlled by
the library. This means that the `await async_*` call in the test case
is enqueuing the logging event, and then the `time.sleep(0.5)` in the
other task blocks the entire event loop for the full half second, and then the
background tasks can complete. We assert `< timestamp - 0.4` to account
for any variance in the exact time that different coroutines finish in
the background.

Aside from the fix to the async logging functions, this commit also
includes a change to the type annotations for the internal
`_async_logging_event` function. This change is necessary to get the
signature of the internal `_{level}` functions to check correctly.
Python's `Callable` type doesn't support keyword only arguments, so to
get MyPy to check it, we have to use a protocol type. This shouldn't
have any runtime cost since we haven't annotated it with the `@runtime`
protocol decorator.