Make `err` Argument to Logging Functions Keyword-Only
I decided to change the optional `err` argument to both the sync and
async logging functions to be keyword-only.

This is the style used in all of the examples, and it is the style I
prefer for this kind of argument because it forces the code at all
callsites to be consistent, and it improves readability for people who
are not familiar with the api of the library and are reading someone
else's source code.

This could be considered a stylistic difference, but I use keyword-only
interfaces to keep myself from introducing inconsistencies in my code
over time, and I think it is worthwhile to enforce this at the interface
level.

This constitutes a breaking change, but the library was released three
days ago, and according to the pypa and github stats, no one is using it
yet, so I am going to introduce this change with a minor version update
instead of a major version since releasing v2 for something like this is
stupid, but I don't want to maintain the positional version of the APIs
until there is enough change to warrant a v2 release.