Rebuild Docs Site for v1.1.0
I rebuilt the docs for the next release. I also found a bug in the
documentation/serve.py script. This script is one that I use in most of
my projects, but I normally have a `docs` directory that contains the
documentation source and a `site` subdir that contains the built site,
so the serve.py script serves from the `site` subdir in the same
directory in most of my projects. GH Actions serves the built docs site
from the top-level `docs` directory, so when using the GH Actions
deployment to GH pages, I have to use a separate directory for the doc
source. I forgot to update the script to accomodate this when I was
setting up this repo.