Fix Race Condition in Async Logging Cleanup
The async queue for a logger needs to be flushed when calling
`deregister_async_logger` in order to ensure no log messages are lost.

This is done by cancelling the task that logs events from the queue and
then processing each event in the queue until it is empty.

I had put the logic for the cleanup into the except block inside the
coroutine for processing the enqueued logging events because a cancelled
task will raise the `asyncio.CancelledError` in the wrapped coroutine,
so by catching that exception in the coroutine itself, we keep all of
the log processing code very close together. The problem is that the
exception is raised in the coroutine on the next iteration of the event
loop, so there is a race condition depending on how the calls to
the async logging functions and the call to deregister the async logger
are nested in different tasks.

I have moved the code for flushing the logging events into the top-level
`asyncio.CancelledError` handler in the `deregister_async_logger`
function so that it will be processed when the task is cancelled
regardless of how tasks get nested.