Add Function to Deregister all Async Loggers
I added test cases for the error conditions when (de)registering async
loggers, and I realized that I needed to add a helper function to
cleanly deregister all currently registered async loggers without a
reference to the loggers themselves at the callsite. This is helpful for
testing, but it's also necessary to simplify resource cleanup on
application shutdown.

I added the `deregister_all_async_loggers` function, which takes no
arguments and loops through the `_QUEUES` registry to cleanly deregister
each logger.

This function should be used to cleanup async loggers in an
application's shutdown routine.