Add Explicit Error for `register_async_logger` Outside Running Loop
the `register_async_logger` function requires a running event loop even
though it is a sync function. This is because it creates an asyncio task
to run in the background, so even though the task is created
synchronously, the event loop can't schedule it if there isn't a loop
running.

The errors that this would give were obscure since it would bubble up
from some nested coroutine not being awaited. I have added an explicit
check for a running loop in the `register_async_logger` function, so it
will now raise a RuntimeError with a clear message indicating exactly
what the problem is. This should reduce the debugging burden on callers.

I also added a test case to check for this specific error condition.