Remove `load_env` Function
I am removing the `load_env` function from the library.

I added this function on a whim thinking that it would be helpful for
testing setting up the library in a project that doesn't normally use
.env files, but the `set_env` function covers this use case, and
projects that do use .env files will already have env loading in place,
so the `load_env` function is simply extra maintenance burden for no
real benefit.

This is another breaking change, but like the change to the `err`
arguments in 2ac3c50, I am releasing this as a minor version change
since there are no users of the library yet. I should not have released
this library as 1.0 initially. Lesson learned.