Add Test for RuntimeError in `make_logger`
I added test cases for the error condition in the `make_logger` function
when the environment does not have the needed vars.

I also made a slight change to the error checking so that it would raise
if the `DEFAULT_GRAMMLOG_LEVEL` env var was missing regardless of the
`GRAMMLOG_LEVEL` var. This makes the logic simpler and enforces the
semantics of the `GRAMMLOG_LEVEL` var as an override of the
`DEFAULT_GRAMMLOG_LEVEL` var and not a var that has meaning on its own.