Fix Bug in Default Logger Test Fixture
The `fixt_default_logger` test fixture truncates the corresponding log
file to ensure that each test has a clean state. This was causing the
tests to faile if the `logs` directory was not created before calling
the fixture function. The `make_logger` function creates the directory
if it doesn't exist yet, so I have moved the call to `make_logger` above
the file trucation in the `fixt_default_logger` function to fix this
error in the test suite.