Add Implicit Queue Registration
This change removes the exception from the `async_*` logging functions
when the logger is not already registered to an async queue.

This makes situations like unit tests that isolate event loops much
less error prone when using the library without affecting the common
case of a single-loop application much.

The `register_async_logger` and `deregister_async_logger` functions are
still available for users who need explicit control over the queue size
or exactly when a queue gets registered to which event loop, but the
most common use-case is more convenient now.